+++ projects/pastanaga/src/lib/calendar/calendar.module.ts
@@ -6,6 +6,10 @@ import { TranslateModule } from '../translate/translate.module';
 import { CalendarComponent } from './calendar.component';
 import { DatePickerComponent } from './date-picker.component';
 import { PopupModule } from '../popup/popup.module';
+import { DateInputComponent } from './date-input.component';
+import { TextFieldModule } from '../textfield/textfield.module';
+import { FormsModule } from '@angular/forms';
+
 
 @NgModule({
     imports: [
@@ -14,9 +18,11 @@ import { PopupModule } from '../popup/popup.module';
         PopupModule,
         TooltipModule,
         TranslateModule,
+        TextFieldModule,
+        FormsModule,
     ],
-    exports: [CalendarComponent, DatePickerComponent],
-    declarations: [CalendarComponent, DatePickerComponent],
+    exports: [CalendarComponent, DatePickerComponent, DateInputComponent],
+    declarations: [CalendarComponent, DatePickerComponent, DateInputComponent],
 })
 export class CalendarModule {
 }